+++ SQL/PerfMetric.sql
@@ -8,7 +8,7 @@ CREATE TABLE [dbo].[PerfMetric]
 (
    [id] INT NOT NULL IDENTITY PRIMARY KEY, -- Primary Key column
    [reportId] INT NOT NULL,
-   [score] NUMERIC(3,2) NOT NULL,
+   [score] INT NOT NULL,
    [displayValue] NVARCHAR(50) NOT NUll,
    [numericValue] NUMERIC(10,3) NOT NULL,
    [title] NVARCHAR(50) NOT NUll,
@@ -23,4 +23,6 @@ GO
 SELECT * FROM [dbo].[PerfMetric]
 
 GO
+
+
